fix: Line에 만들어야 할 테스트를 Ladder에 만들어서 이를 옮김
